Kernel/Graphics: Export Intel definitions to a header file
