+++ Kernel/Graphics/Intel/DisplayConnectorGroup.h
@@ -9,6 +9,7 @@
 #include <AK/RefPtr.h>
 #include <AK/Try.h>
 #include <Kernel/Graphics/Console/GenericFramebufferConsole.h>
+#include <Kernel/Graphics/Intel/Definitions.h>
 #include <Kernel/Graphics/Intel/GMBusConnector.h>
 #include <Kernel/Graphics/Intel/NativeDisplayConnector.h>
 #include <Kernel/Library/LockRefPtr.h>
@@ -16,48 +17,6 @@
 #include <LibEDID/EDID.h>
 
 namespace Kernel {
-
-namespace IntelGraphics {
-
-enum class RegisterIndex {
-    PipeAConf = 0x70008,
-    PipeBConf = 0x71008,
-    GMBusData = 0x510C,
-    GMBusStatus = 0x5108,
-    GMBusCommand = 0x5104,
-    GMBusClock = 0x5100,
-    DisplayPlaneAControl = 0x70180,
-    DisplayPlaneBControl = 0x71180,
-    DisplayPlaneALinearOffset = 0x70184,
-    DisplayPlaneAStride = 0x70188,
-    DisplayPlaneASurface = 0x7019C,
-    DPLLDivisorA0 = 0x6040,
-    DPLLDivisorA1 = 0x6044,
-    DPLLControlA = 0x6014,
-    DPLLControlB = 0x6018,
-    DPLLMultiplierA = 0x601C,
-    HTotalA = 0x60000,
-    HBlankA = 0x60004,
-    HSyncA = 0x60008,
-    VTotalA = 0x6000C,
-    VBlankA = 0x60010,
-    VSyncA = 0x60014,
-    PipeASource = 0x6001C,
-    AnalogDisplayPort = 0x61100,
-    VGADisplayPlaneControl = 0x71400,
-};
-
-struct PLLSettings;
-
-struct PLLParameterLimit {
-    size_t min, max;
-};
-
-struct PLLMaxSettings {
-    PLLParameterLimit dot_clock, vco, n, m, m1, m2, p, p1, p2;
-};
-
-}
 
 class IntelDisplayConnectorGroup : public RefCounted<IntelDisplayConnectorGroup> {
     friend class IntelNativeGraphicsAdapter;
